Renamed AmqpMessageSender to MessageSender, and refactored accordingly
